Party list is now shown on new load
Fixes #24 on github.

The issue was with how the render() function was called. In the way the
code was set up the render() funciton was called on init (and an empty
collection) - any further sync event would just execute the return function
of the function. I've encapsulated the render() call in a function() {
this.render() } - now it works.